Fixed db deletion only when it exists, for testing purposes

+++ tests/test_1_user.py
@@ -2,7 +2,9 @@ import sys
 import os
 os.chdir('src/')
 sys.path.append(os.getcwd())
-os.remove(os.getcwd()+'/brainless.db')
+
+if os.path.exists(os.getcwd()+'/brainless.db'):
+    os.remove(os.getcwd()+'/brainless.db')
 
 import unittest
 from configuration import db
